Complete refactor of handling split Tx/Rx.
* Change UartRx/UartTx to a unified Uart struct (with
  capability-dictating type parameter)
* Add Capability, Transmit, Receive, Simplex traits
* Implement AsMut<Reconfig> for (&mut Uart<C, RxDuplex>, &mut Uart<C,
  TxDuplex>) to allow reconfiguring without needing to move Rx/Tx halves
  into a Uart<C, Duplex>
* Split some code into submodules
* Add RegisterInterface trait to help with some register accesses
* Update docs
